Move a method, remove classes.
