+++ core-java/src/main/java/org/spine3/server/CommandStore.java
@@ -24,7 +24,6 @@ import com.google.protobuf.Any;
 import com.google.protobuf.Message;
 import org.spine3.base.CommandRequest;
 import org.spine3.protobuf.Messages;
-import org.spine3.server.aggregate.AggregateCommand;
 import org.spine3.server.aggregate.AggregateId;
 import org.spine3.server.storage.CommandStorage;
 
@@ -49,7 +48,7 @@ public class CommandStore {
     public void store(CommandRequest request) {
         final Any any = request.getCommand();
         final Message command = Messages.fromAny(any);
-        final AggregateId aggregateId = AggregateCommand.getAggregateId(command);
+        final AggregateId aggregateId = AggregateId.getAggregateId(command);
         storage.store(aggregateId, request);
     }
